requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
PyPDF2
python-docx
python-dotenv
beautifulsoup4
lxml
huggingface-hub
